factor out the config
